--- .env
@@ -1,1 +1,0 @@
-INLINE_RUNTIME_CHUNK=false
