+++ package.json
@@ -17,7 +17,6 @@
     "upgrade": "npx projen upgrade",
     "default": "npx projen default",
     "watch": "npx projen watch",
-    "package": "npx projen package",
     "eslint": "npx projen eslint",
     "dev": "npx projen dev",
     "export": "npx projen export",
@@ -32,7 +31,6 @@
     "next:start": "npx projen next:start",
     "lint": "npx projen lint",
     "tsc": "npx projen tsc",
-    "dev:fetch-assemblies": "npx projen dev:fetch-assemblies",
     "release": "npx projen release",
     "projen": "npx projen"
   },
@@ -100,6 +98,5 @@
   },
   "license": "Apache-2.0",
   "version": "0.0.0",
-  "proxy": "https://construct-hub-testing.dev-tools.aws.dev/",
   "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
 }
