+++ src/pages/faq.ts
