+++ src/pages/packages/[name]/[scopedName]/index.ts
