+++ src/pages/packages/[name]/[scopedName]/v/[version].ts
