+++ src/pages/packages/[name]/index.ts
