+++ src/pages/packages/l/[lang]/[name]/[scopedName]/v/[version]/index.ts
