+++ src/pages/packages/l/[lang]/[name]/[scopedName]/v/[version]/s/[submodule]/index.ts
