+++ src/pages/packages/l/[lang]/[name]/v/[version]/index.ts
