+++ src/pages/packages/l/[lang]/[name]/v/[version]/s/[submodule]/index.ts
