+++ src/pages/terms.ts
