+++ tsconfig.json
@@ -44,8 +44,7 @@
   "include": [
     "pages/**/*.ts",
     "src/**/*.ts",
-    "src/**/*.tsx",
-    "pages/**/*.tsx"
+    "src/**/*.tsx"
   ],
   "exclude": [],
   "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
